requirements.txt
flask
requests
psutil
opencv-python>=4.8.0
numpy>=1.24.0
mss>=9.0.0
pyautogui>=0.9.50
Pillow>=10.0.0
pywin32>=305
PyYAML>=6.0
